Add spring common extension
fixes #538
