+++ extensions/sql/deployment/src/main/java/org/apache/camel/quarkus/component/sql/deployment/SqlProcessor.java
@@ -20,16 +20,11 @@ import java.sql.Types;
 
 import org.apache.camel.quarkus.component.sql.CamelSqlConfig;
 
-import io.quarkus.deployment.GeneratedClassGizmoAdaptor;
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
-import io.quarkus.deployment.builditem.GeneratedClassBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.NativeImageResourceBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.ReflectiveClassBuildItem;
-import io.quarkus.deployment.pkg.steps.NativeBuild;
-import io.quarkus.gizmo.ClassCreator;
-import io.quarkus.gizmo.ClassOutput;
 
 class SqlProcessor {
 
@@ -55,19 +50,5 @@ class SqlProcessor {
                 .stream()
                 .map(scriptFile -> new NativeImageResourceBuildItem(scriptFile.replace("classpath:", "")))
                 .forEach(nativeImage::produce);
-    }
-
-    @BuildStep(onlyIf = NativeBuild.class)
-    void generateKParameterClass(BuildProducer<GeneratedClassBuildItem> generatedClass) {
-        // TODO: Investigate removing this. See https://github.com/apache/camel-quarkus/issues/534
-        // The native image build fails with a NoClassDefFoundError without this. Possibly similar to https://github.com/oracle/graal/issues/656.
-        ClassOutput classOutput = new GeneratedClassGizmoAdaptor(generatedClass, false);
-        ClassCreator.builder()
-                .className("kotlin.reflect.KParameter")
-                .classOutput(classOutput)
-                .setFinal(true)
-                .superClass(Object.class)
-                .build()
-                .close();
     }
 }
