+++ extensions/support/spring/runtime/src/main/java/org/apache/camel/quarkus/support/spring/graal/SpringSubstitutions.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.quarkus.component.sql.graal;
+package org.apache.camel.quarkus.support.spring.graal;
 
 import java.io.IOException;
 import java.lang.reflect.Constructor;
