+++ integration-tests/sql/src/main/java/org/apache/camel/quarkus/component/sql/it/SqlResource.java
@@ -95,6 +95,7 @@ public class SqlResource {
                 .build();
     }
 
+    @SuppressWarnings("unchecked")
     @Path("/storedproc")
     @GET
     @Produces(MediaType.TEXT_PLAIN)
