+++ pom.xml
@@ -50,6 +50,7 @@
         <snakeyaml.version>1.25</snakeyaml.version>
         <xalan.version>2.7.2</xalan.version>
         <kotlin.version>1.3.21</kotlin.version>
+        <spring.version>5.2.1.RELEASE</spring.version>
 
         <maven-compiler-plugin.version>3.8.0</maven-compiler-plugin.version>
         <maven.compiler.target>1.8</maven.compiler.target>
