+++ poms/bom-deployment/pom.xml
@@ -376,6 +376,11 @@
             </dependency>
             <dependency>
                 <groupId>org.apache.camel.quarkus</groupId>
+                <artifactId>camel-quarkus-support-spring-deployment</artifactId>
+                <version>${camel-quarkus.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.apache.camel.quarkus</groupId>
                 <artifactId>camel-quarkus-support-xalan-deployment</artifactId>
                 <version>${camel-quarkus.version}</version>
             </dependency>
